Add support for Codabar format
This adds support for the Codabar format (https://en.wikipedia.org/wiki/Codabar, http://www.barcodeisland.com/codabar.phtml). Codabar supports encoding 16 different characters (0-9, -$:/.+), plus four additional start/stop characters (A-D).

The start and stop characters do not appear in the body of a Codabar string, and are stripped from the `text` propery returned by the encoder. If no start/stop characters are passed to the encoder, "A" is assumed for each.