Added support for CODE93FullASCII
Add support for encoding the full ASCII table in CODE93, using two
symbol encodings per character for the extended set.

CODE93 no longer automatically converts the input data to upper case,
use CODE93FullASCII if lower case inputs are needed.

The symbol for the stop/end sequence is changed to '\xff', as '*' is
a valid input character in CODE93FullASCII.